+++ app/src/main/java/com/chico/myhomebookkeeping/ui/firstLaunch/FirstLaunchSetupItem.kt
@@ -1,5 +1,7 @@
 package com.chico.myhomebookkeeping.ui.firstLaunch
 
+import com.chico.myhomebookkeeping.enums.icon.names.CategoryIconNames
+
 data class FirstLaunchSetupItem(
     val img: Int,
     val name: String
@@ -9,8 +11,10 @@ data class FirstLaunchCategoryGroupItem(
     val parentName: String,
     val parentNameRu: String?,
     val parentNamePl: String?,
+    val parentIcon: CategoryIconNames,
     val isIncome: Boolean,
     val subcategories: List<String>,
     val subcategoriesRu: List<String>,
-    val subcategoriesPl: List<String>
+    val subcategoriesPl: List<String>,
+    val subcategoryIcons: List<CategoryIconNames>
 )